+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.2
+
+- Improved expiration (no timers)
+- `Refresher` -> `Callable`
+
 ## 0.0.1
 
 - Initial version.